requirements.txt
# requirements
flask
scikit-learn
numpy
pymupdf
python-docx
tqdm
sentence-transformers
nltk
chromadb
openai
python-dotenv
gunicorn
azure-storage-blob